chore: restructure password-reset directory
