+++ backend/internal/handlers/auth/forgot_pass/forgot_pass.go
@@ -1,0 +1,1 @@
+package forgot_pass
