+++ backend/internal/handlers/auth/forgot_pass/service.go
@@ -1,0 +1,1 @@
+package forgot_pass
